pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=1.0.0",
    "phidata>=2.7.10",
    "streamlit>=1.53.1",
]
